updated data

+++ src/main/java/com/training/demo/repository/CoinManagementRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.training.demo.model.CoinManagement;
 import com.training.demo.model.User;
 
-public interface CoinManagementRepository extends JpaRepository<CoinManagement, Long> {
+public interface CoinManagementRepository extends JpaRepository<CoinManagement, Integer> {
 
-	public CoinManagement findOneByCoinId(Integer coinId);
+	public CoinManagement findOneByCoinId(long coinId);
 
 	public CoinManagement findByCoinName(String coinName);
 
@@ -17,6 +17,7 @@ public interface CoinManagementRepository extends JpaRepository<CoinManagement, 
 
 	public CoinManagement findOneByCoinName(String coinName);
 
+//	public void deleteByCoinId(Integer coinId);
 	//public CoinManagement findByCoinNameAndSymbol(String coinName, String sm);
 
 }
